--- src/components/molecules/EventList.js
@@ -1,14 +1,0 @@
-import React from 'react';
-import EventCard from '../atoms/EventCard';
-
-const EventList = ({ events }) => {
-    return (
-        <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-6">
-            {events.map((event, index) => (
-                <EventCard key={index} event={event} />
-            ))}
-        </div>
-    );
-};
-
-export default EventList;
